Moved content and session folder to usecase
