+++ server/tests/content.go
@@ -1,9 +1,9 @@
 package tests
 
 import (
-	"choco/server/internals/content"
 	"choco/server/internals/models"
-	"choco/server/internals/session"
+	"choco/server/internals/usecase/content"
+	"choco/server/internals/usecase/session"
 	"testing"
 )
